Move ID generation to Trace module
